Feat (Playtest): Change input delay from 1s to 0.5s

+++ Assets/_Scripts/Objects/ObjectSpawner.cs
@@ -40,7 +40,7 @@ public class ObjectSpawner : MonoBehaviour
     [SerializeField] private int upperLimit;
 
     [Header("Input Delay")]
-    [SerializeField] private int InputDelayTime;
+    [SerializeField] private float InputDelayTime;
 
     [Header("Game Objects")]
     [SerializeField] private GameObject correctParticles;
